Added stats and other stuff
Added stats and changed datatype from int to float to represent the
values (for future stat modifiers, percentage and such).
Made changes accordingly to this.
Removed InCombat from CombatSystem, the game state will be handled by
the GameManager.

+++ Assets/Scripts/IDamageable.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 
 public interface IDamagable
 {
-    int Health { get; }
-    IEnumerator Damage(int damage, Vector2 direction);
+    float Health { get; }
+    IEnumerator Damage(float damage, Vector2 direction);
     IEnumerator DamageEffect();
     void Kill();
 }
